requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
flask>=2.0.0 